opal-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.4.0
pydantic[email]==2.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0
pandas==2.2.0
scikit-learn==1.4.0
joblib==1.3.2
numpy==1.26.4
pytest==8.0.0
pytest-asyncio==0.23.5
google-generativeai==0.3.2
pydantic-settings==2.1.0
